+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [11.4.0] - 2023-01-13
 - Disable Gravity Forms update messages and background updates
 - Flush Kinsta cache rather than WP-Rocket when present
-- Don't force mailers
+- Don't force mailers in any environment without proper definitions present
+- Don't enable SMTP mailers in non-production environments unless specifically declared
 
 ## [11.3.1] - 2022-12-15
 - Define Kinsta mu-plugin URL